requirements.txt
django==4.2.1
psycopg2==2.9.6
python-dotenv==1.0.0
bs4
requests==2.30.0
Pillow==9.5.0
django-crispy-forms==2.0
django-encrypted-model-fields==0.6.5